requirements.txt
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
Flask-Migrate==3.1.0
Flask-JWT-Extended==4.3.1
psycopg2-binary==2.9.1
python-dotenv==0.19.0
cryptography==3.4.7
bcrypt==3.2.0
requests==2.26.0
gunicorn==20.1.0
pytest==6.2.5
black==21.7b0
flake8==3.9.2 